WEB/Flask Server/requirements.txt
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
Pillow==7.2.0
six==1.15.0
Werkzeug==1.0.1
wincertstore==0.2
